v0.2 Kindling: Frontend shell + Campfire chat
Sprint 2 implementation complete:
- Frontend: Vite + React 19 + TypeScript strict + Tailwind v4 (27 files)
- Subtle Warmth design system (dark/light, @fontsource, pillow buttons)
- Campfire: real-time SSE chat, 4-stage CSS fade, optimistic send
- Auth: AuthProvider, token auto-refresh, SSE reconnect resync (SEC-006)
- Hooks: useMessages, usePresence, useReconnect, useTimeSync
- Components: MessageBubble, MumblingIndicator, PresenceBar, Shell
- Code-splitting via React.lazy (K-024)
- PocketBase serves SPA from pb_public/
- Backend: added CORS, rate limiting, input sanitization hooks
- Build: tsc clean, 91KB gzipped (under 150KB budget)

Research & docs:
- R-008: CSS animation performance (compositor-thread safe at 200+)
- R-009: Marketing prep, Reddit post draft
- Sprint 2 spec, LAN setup guide
- ROADMAP, CHRONICLE, BACKLOG updated

+++ backend/hooks/message_gc.go
@@ -25,7 +25,7 @@ func RegisterMessageGC(app *pocketbase.PocketBase) {
 		if affected, _ := res.RowsAffected(); affected > 0 {
 			app.Logger().Info("message GC sweep", "deleted", affected)
 			// Increment Prometheus counter (tracked in metrics.go)
-			gcDeletedTotal.Add(float64(affected))
+			gcDeletedTotal.Add(affected)
 		}
 	})
 }
